get busstops data from Firebase & apply MVVM
